requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
duckdb>=0.10.0
pandas>=2.2.0
pyarrow>=16.0.0
scipy>=1.13.0
anthropic>=0.28.0
Pillow>=10.3.0
python-dotenv>=1.0.0
pydantic>=2.7.0
streamlit>=1.35.0
numpy>=1.26.0
